fix(handler.spec_field): can now accept false without losing laziness
lze is meant to work such that if you use a field from a spec,
and the handler is marked to affect laziness, then that should make it
lazy.

Before, a handler could not accept false as a value via its spec
field without the plugin becoming eager.

Now, the handler will affect lazy value in all cases
where the handler has set_lazy = true
and the field is used (is not nil).